+++ app/src/main/kotlin/com/jjdev/equi/MainActivity.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
 import com.jjdev.equi.core.ui.theme.EquiTheme
+import com.jjdev.equi.dashboard.presentation.DashboardScreen
 
 class MainActivity : ComponentActivity() {
