+++ app/src/main/kotlin/com/jjdev/equi/dashboard/domain/model/Investment.kt
@@ -1,4 +1,4 @@
-package com.jjdev.equi
+package com.jjdev.equi.dashboard.domain.model
 
 data class Investment(
     val ticker: String,
