+++ app/src/main/kotlin/com/jjdev/equi/dashboard/presentation/model/DashboardScreenViewState.kt
@@ -1,4 +1,4 @@
-package com.jjdev.equi
+package com.jjdev.equi.dashboard.presentation.model
 
 data class DashboardScreenViewState(
     val pieChartData: Any? = null,
