+++ app/src/test/kotlin/com/jjdev/equi/RebalanceUseCaseTest.kt
@@ -2,6 +2,8 @@ package com.jjdev.equi
 
 import com.jjdev.equi.core.base.domain.onError
 import com.jjdev.equi.core.base.domain.onSuccess
+import com.jjdev.equi.dashboard.domain.RebalanceUseCase
+import com.jjdev.equi.dashboard.domain.model.Investment
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
